requirements.txt
google-cloud-language
flask
azure-cognitiveservices-search-newssearch